refactor(query_api): convert to a directory; lift the 1.3k-line test module to its own file
index/query_api.rs was 3900 lines (god-module #1 by repo_brief: 146 symbols,
fan_in 335). Convert it to index/query_api/ and move the oracle_surfacing_tests
module into its own oracle_surfacing_tests.rs (1308 lines). Pure move — no
behavior change; production file drops to ~2591 lines. First step of splitting
the impl block by job.
